02-02 -> 02-06
Creation des repositories, services et controllers pour bike, motorbike, campingcar & utility.
Creation des DTO respectifs pour les Request, PatchRequest, CustomerResponse et AdminResponse pour bike, motorbike, campingcar & utility.

Insertion de méthodes de vérification métier dans vehicle, fourwheeled, cars, campingcar, utility, bike & motorbike.

Les méthodes getAll, getById, Create, Patch, Delete et filtre par active/outoffleet sont toutes fonctionnelles.

+++ src/main/java/com/accenture/locationvoitures/service/dto/request/vehicle/CampingCarRequestDto.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 public final class CampingCarRequestDto extends FourWheeledRequestDto {
     @NotNull
     @Min(0)
-    private Integer ptac;
+    private Double ptac;
     @NotNull
     @Min(0)
     private Double height;
