+++ src/main/java/com/accenture/locationvoitures/service/dto/response/admin/vehicle/CampingCarAdminResponseDto.java
@@ -17,7 +17,7 @@ public record CampingCarAdminResponseDto(
         String transmission,
         Boolean airConditioning,
 
-        Integer ptac,
+        Double ptac,
         Double height,
         CampingCarEquipmentResponseDto campingCarEquipment,
